+++ src/usermanager.cpp
@@ -170,7 +170,7 @@ void UserManager::AddUser(int socket, ListenSocket* via, irc::sockets::sockaddrs
 
 	if (this->local_users.size() > ServerInstance->Config->SoftLimit)
 	{
-		ServerInstance->SNO.WriteToSnoMask('a', "Warning: softlimit value has been reached: %d clients", ServerInstance->Config->SoftLimit);
+		ServerInstance->SNO.WriteToSnoMask('a', "Warning: softlimit value has been reached: %lu clients", ServerInstance->Config->SoftLimit);
 		this->QuitUser(New,"No more connections allowed");
 		return;
 	}
@@ -375,7 +375,7 @@ void UserManager::DoBackgroundUserStuff()
 
 		if (curr->CommandFloodPenalty || curr->eh.GetSendQSize())
 		{
-			unsigned int rate = curr->GetClass()->GetCommandRate();
+			unsigned long rate = curr->GetClass()->GetCommandRate();
 			if (curr->CommandFloodPenalty > rate)
 				curr->CommandFloodPenalty -= rate;
 			else
